pyproject.toml
[tool.poetry]
name = "carto-waze-lambda"
version = "0.1.0"
description = "CARTO Waze Lambda Connector"
authors = ["CARTO Professional Services"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "~2.25.1"
carto = "~1.11.2"
google-cloud-bigquery = "~2.10.0"

[tool.poetry.dev-dependencies]
pylint = "^2.7.2"
pylint-quotes = "^0.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
